third_party/WebKit/LayoutTests/fast/dom/Selection/collapseToX-empty-selection-expected.txt

Test that collapseToStart() and collapseToEnd() throw INVALID_STATE_ERR if no selection is made.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sel.collapseToStart() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS sel.collapseToEnd() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS sel.collapseToStart() is undefined
PASS sel.collapseToEnd() is undefined
PASS successfullyParsed is true

TEST COMPLETE
